+++ src/Controllers/Product/ProductGermanMarketFieldsController.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace JtlWooCommerceConnector\Controllers\Product;
 
-use InvalidArgumentException;
 use Jtl\Connector\Core\Model\Product as ProductModel;
+use Jtl\Connector\Core\Model\TranslatableAttribute;
 use JtlWooCommerceConnector\Controllers\AbstractBaseController;
 use PhpUnitsOfMeasure\Exception\NonNumericValue;
 use PhpUnitsOfMeasure\Exception\NonStringUnitName;
@@ -21,7 +21,7 @@ class ProductGermanMarketFieldsController extends AbstractBaseController
      * @param ProductModel $product
      * @param WC_Product   $wcProduct
      * @return void
-     * @throws InvalidArgumentException
+     * @throws \InvalidArgumentException
      */
     public function pullData(ProductModel &$product, WC_Product $wcProduct): void
     {
@@ -32,7 +32,7 @@ class ProductGermanMarketFieldsController extends AbstractBaseController
      * @param ProductModel $product
      * @param WC_Product   $wcProduct
      * @return void
-     * @throws InvalidArgumentException
+     * @throws \InvalidArgumentException
      */
     private function setBasePriceProperties(ProductModel $product, WC_Product $wcProduct): void
     {
@@ -359,7 +359,7 @@ class ProductGermanMarketFieldsController extends AbstractBaseController
      * @param ProductModel          $product
      * @param array<string, string> $metaKeys
      * @return void
-     * @throws \http\Exception\InvalidArgumentException
+     * @throws \InvalidArgumentException
      */
     private function clearPPU(ProductModel $product, array $metaKeys): void
     {
@@ -412,8 +412,8 @@ class ProductGermanMarketFieldsController extends AbstractBaseController
 
         [$gpsrManufacturerAddress, $gpsrResponsibleAddress] = $this->createManufacturerAndResponsibleStrings($product);
 
-        \update_post_meta($postId, '_german_market_gpsr_manufacturer', $gpsrManufacturerAddress);
-        \update_post_meta($postId, '_german_market_gpsr_responsible_person', $gpsrResponsibleAddress);
+        \update_post_meta((int)$postId, '_german_market_gpsr_manufacturer', $gpsrManufacturerAddress);
+        \update_post_meta((int)$postId, '_german_market_gpsr_responsible_person', $gpsrResponsibleAddress);
     }
 
     /**
@@ -511,6 +511,7 @@ class ProductGermanMarketFieldsController extends AbstractBaseController
             }
         }
 
+        /** @var array<string, string> $manufacturerData */
         $gpsrManufacturerAddress = $manufacturerData['name'] . "\n"
             . $manufacturerData['street'] . ' ' . $manufacturerData['housenumber'] . "\n"
             . $manufacturerData['postalcode'] . ' ' . $manufacturerData['city'] . "\n"
@@ -518,6 +519,7 @@ class ProductGermanMarketFieldsController extends AbstractBaseController
             . $manufacturerData['email'] . "\n"
             . $manufacturerData['homepage'];
 
+        /** @var array<string, string> $responsiblePersonData */
         $gpsrResponsibleAddress = $responsiblePersonData['name'] . "\n"
             . $responsiblePersonData['street'] . ' ' . $responsiblePersonData['housenumber'] . "\n"
             . $responsiblePersonData['postalcode'] . ' ' . $responsiblePersonData['city'] . "\n"