+++ src/Controllers/Product/ProductSpecialPriceController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace JtlWooCommerceConnector\Controllers\Product;
 
-use InvalidArgumentException;
+use Jtl\Connector\Core\Model\CustomerGroup;
 use Jtl\Connector\Core\Model\CustomerGroup as CustomerGroupModel;
 use Jtl\Connector\Core\Model\CustomerGroupI18n as CustomerGroupI18nModel;
 use Jtl\Connector\Core\Model\Identity;
@@ -27,7 +27,7 @@ class ProductSpecialPriceController extends AbstractBaseController
      * @param WC_Product   $product
      * @param ProductModel $model
      * @return ProductSpecialPriceModel[]
-     * @throws InvalidArgumentException
+     * @throws \InvalidArgumentException
      * @throws \Exception
      */
     public function pullData(WC_Product $product, ProductModel $model): array
@@ -129,7 +129,7 @@ class ProductSpecialPriceController extends AbstractBaseController
      * @param string     $priceNet
      * @param WC_Product $product
      * @return float
-     * @throws \http\Exception\InvalidArgumentException
+     * @throws \InvalidArgumentException
      */
     protected function getPriceNet(string $priceNet, WC_Product $product): float
     {
@@ -150,7 +150,7 @@ class ProductSpecialPriceController extends AbstractBaseController
      * @param WC_Product   $wcProduct
      * @param string       $productType
      * @return void
-     * @throws InvalidArgumentException
+     * @throws \InvalidArgumentException
      * @throws \Exception
      */
     public function pushData(ProductModel $product, WC_Product $wcProduct, string $productType): void
@@ -210,12 +210,12 @@ class ProductSpecialPriceController extends AbstractBaseController
     }
 
     /**
-     * @param ProductModel $product
+     * @param ProductModel          $product
      * @param ProductSpecialPrice[] $specialPrices
-     * @param string $productId
-     * @param Identity $masterProductId
-     * @param string $productType
-     * @param int $pd
+     * @param string                $productId
+     * @param Identity              $masterProductId
+     * @param string                $productType
+     * @param int                   $pd
      * @return void
      * @throws \Psr\Log\InvalidArgumentException
      */
@@ -235,6 +235,7 @@ class ProductSpecialPriceController extends AbstractBaseController
                 if ($specialPrice->getConsiderDateLimit()) {
                     $dateTo = \is_null($end = $specialPrice->getActiveUntilDate())
                         ? null
+                        // @phpstan-ignore-next-line
                         : $end->setTime(
                             Date::LAST_HOUR,
                             Date::LAST_MINUTE,
@@ -317,7 +318,7 @@ class ProductSpecialPriceController extends AbstractBaseController
                     }
                 } elseif (\is_int((int)$endpoint)) {
                     if ($productType !== ProductController::TYPE_PARENT) {
-                        $customerGroup = \get_post($endpoint);
+                        $customerGroup = \get_post((int)$endpoint);
                         $priceMetaKey  = null;
 
                         if (!$customerGroup instanceof \WP_Post) {
@@ -593,25 +594,27 @@ class ProductSpecialPriceController extends AbstractBaseController
                             }
                         } else {
                             /** @var false|string $regularPrice */
-                            $regularPrice = \get_post_meta($productId, $regularPriceMetaKey, true);
+                            $regularPrice = \get_post_meta((int)$productId, $regularPriceMetaKey, true);
                             if (!\is_bool($regularPrice)) {
                                 \update_post_meta(
                                     (int)$productId,
-                                    $priceMetaKey,
+                                    $priceMetaKey ?? '_price',
                                     \wc_format_decimal((float)$regularPrice, $pd),
-                                    \get_post_meta((int)$productId, $priceMetaKey, true)
+                                    \get_post_meta((int)$productId, $priceMetaKey ?? '_price', true)
                                 );
                             }
                         }
                     }
-                } else {
-                    continue;
                 }
             }
         }
     }
 
-    public function setCustomerGroupNames($customerGroups): void
+    /**
+     * @param CustomerGroup[] $customerGroups
+     * @return void
+     */
+    public function setCustomerGroupNames(array $customerGroups): void
     {
         if (
             SupportedPlugins::comparePluginVersion(
@@ -633,15 +636,15 @@ class ProductSpecialPriceController extends AbstractBaseController
     }
 
     /**
-     * @param $customerGroups
-     * @param string $productId
-     * @param Identity $masterProductId
-     * @param string $productType
-     * @param int $pd
+     * @param CustomerGroup[] $customerGroups
+     * @param string          $productId
+     * @param Identity        $masterProductId
+     * @param string          $productType
+     * @param int             $pd
      * @return void
      */
     public function updateCustomerGroupPostMeta(
-        $customerGroups,
+        array $customerGroups,
         string $productId,
         Identity $masterProductId,
         string $productType,
@@ -759,9 +762,9 @@ class ProductSpecialPriceController extends AbstractBaseController
 
             \update_post_meta(
                 (int)$productId,
-                $priceMetaKey,
+                $priceMetaKey ?? '_price',
                 \wc_format_decimal((float)$regularPrice, $pd),
-                \get_post_meta((int)$productId, $priceMetaKey, true)
+                \get_post_meta((int)$productId, $priceMetaKey ?? '_price', true)
             );
 
             if ($productType === ProductController::TYPE_CHILD) {
@@ -787,8 +790,8 @@ class ProductSpecialPriceController extends AbstractBaseController
 
     /**
      * @param string $productId
-     * @param \WP_Post|null $post
-     * @return string
+     * @param string $postName
+     * @return string|null
      */
     public function setPostMetaKey(
         string $productId,
@@ -809,10 +812,10 @@ class ProductSpecialPriceController extends AbstractBaseController
 
             $metaKeyForCustomerGroupPriceType = $priceMetaKey . '_type';
             \update_post_meta(
-                $productId,
+                (int)$productId,
                 $metaKeyForCustomerGroupPriceType,
                 'fix',
-                \get_post_meta($productId, $metaKeyForCustomerGroupPriceType, true)
+                \get_post_meta((int)$productId, $metaKeyForCustomerGroupPriceType, true)
             );
         }
 
@@ -820,11 +823,11 @@ class ProductSpecialPriceController extends AbstractBaseController
     }
 
     /**
-     * @param string $productId
-     * @param $post
-     * @return array
+     * @param string   $productId
+     * @param \WP_Post $post
+     * @return array<int, string|null>
      */
-    public function setPriceMetaKeysForTypeChild(string $productId, $post): array
+    public function setPriceMetaKeysForTypeChild(string $productId, \WP_Post $post): array
     {
         $COPpriceMetaKey     = null;
         $COPpriceTypeMetaKey = null;